Updating comments and python structure for nexus teams activity worker
