+++ nexus/ui_connector/teams_activity_worker/src/teams_activity_worker/__init__.py
@@ -1,6 +1,5 @@
 """Microsoft Teams Temporal activity worker."""
 
-from .activities import TeamsActivities
+from .platform import TeamsActivities
 
 __all__ = ["TeamsActivities"]
-
