--- nexus/ui_connector/teams_activity_worker/src/teams_activity_worker/__main__.py
@@ -1,4 +1,0 @@
-from .worker import main
-
-main()
-
