+++ nexus/ui_connector/teams_activity_worker/tests/test_config.py
@@ -1,6 +1,6 @@
 import pytest
 
-from teams_activity_worker.config import DEFAULT_SERVICE_URL, Settings
+from teams_activity_worker.platform import DEFAULT_SERVICE_URL, Settings
 
 
 def test_settings_reads_existing_environment_names(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -22,4 +22,3 @@ def test_settings_requires_credentials(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.delenv("MICROSOFT_TENANT_ID", raising=False)
     with pytest.raises(ValueError, match="MICROSOFT_TENANT_ID is required"):
         Settings.from_env()
-
